ported xtest, increased harness, re-ported TA

+++ opentee/TAs/optee_pkcs11_ta/src/pkcs11_helpers.h
@@ -6,14 +6,13 @@
 #ifndef PKCS11_HELPERS_H
 #define PKCS11_HELPERS_H
 
-#include "pkcs11_ta.h"
+#include <pkcs11_ta.h>
 #include <stdint.h>
 #include <stddef.h>
-#include "tee_internal_api.h" // opentee
-#include "tee_logging.h"
+#include <tee_internal_api.h>
 
-#include "pkcs11_attributes.h"
-#include "token_capabilities.h"
+#include <pkcs11_attributes.h>
+#include <token_capabilities.h>
 
 struct pkcs11_object;
 
@@ -68,12 +67,13 @@ bool pkcs2tee_load_attr(TEE_Attribute *tee_ref, uint32_t tee_id,
 			enum pkcs11_attr_id pkcs11_id);
 
 /* Hash and load TEE operation attributes from a PKCS11 object */
+// ! ### enum pkcs11_rc pkcs2tee_load_hashed_attr(TEE_Attribute *tee_ref, uint32_t tee_id, struct pkcs11_object *obj, enum pkcs11_attr_id pkcs11_id, uint32_t tee_algo, void *hash_ptr, uint32_t *hash_size);
 enum pkcs11_rc pkcs2tee_load_hashed_attr(TEE_Attribute *tee_ref,
 					 uint32_t tee_id,
 					 struct pkcs11_object *obj,
 					 enum pkcs11_attr_id pkcs11_id,
 					 uint32_t tee_algo, void *hash_ptr,
-					 uint32_t *hash_size);
+					 size_t *hash_size);
 
 /* Return true if attribute is a boolean, false otherwise */
 static inline bool pkcs11_attr_is_boolean(enum pkcs11_attr_id id)
